added composites

+++ Interfaces/IHeuristic.cs
@@ -3,6 +3,6 @@ namespace BoltFreezer.Interfaces
 {
     public interface IHeuristic
     {
-        int Heuristic(IPlan plan);
+        float Heuristic(IPlan plan);
     }
 }
